Move `store` to root directory; delete empty `app` directory.
