+++ src/store.js
@@ -1,6 +1,6 @@
 import { configureStore } from "@reduxjs/toolkit";
-import cartReducer from "../features/cartSlice";
-import shopReducer from "../features/shopSlice";
+import cartReducer from "./features/cartSlice";
+import shopReducer from "./features/shopSlice";
 
 export const store = configureStore({
   reducer: {
